Account for renaming analytics.ts -> segment-analytics.ts
